Only int64_t

+++ runtime_core/interpretator.cpp
@@ -1,15 +1,15 @@
 #include "interpretator.h"
-#include "isa_opcode_defines.h"
+#include "defines.h"
 
 void Frame::DumpRegs() {
     std::cout << "Frame dump (" << _regs.size() << "):\n";
-    for (uint8_t i = 0; i < _regs.size(); ++i) {
+    for (int64_t i = 0; i < _regs.size(); ++i) {
         std::cout << "REG[" << static_cast<int64_t>(i) << "]: " << _regs[i] << std::endl;
     }
 }
 
 
-int64_t &Interpretator::GetRegRef(const uint64_t n) {
+int64_t &Interpretator::GetRegRef(const int64_t n) {
     if (_current_frame->_n_regs == 0 || n > _current_frame->_n_regs - 1) {
         throw std::runtime_error("Out of bounds in registers.");
     }
@@ -30,7 +30,7 @@ Interpretator::Interpretator() {
 }
 
 
-uint64_t Interpretator::Run(bool verbose) {
+int64_t Interpretator::Run(bool verbose) {
     if (_ptr == nullptr) {
         throw std::runtime_error("Can't Run bytecode, _ptr = nullptr.");
     }
@@ -41,13 +41,13 @@ uint64_t Interpretator::Run(bool verbose) {
 
     try {
         while (*_ptr != OPCODE_EXIT) {
-            uint8_t opcode = *_ptr;
+            int64_t opcode = *_ptr;
             if (opcode > OPCODE_EXIT || DoInstr[opcode] == nullptr) {
                 throw std::runtime_error("Invalid opcode: " + std::to_string(opcode));
             }
 
             if (verbose) {
-                uint8_t *old_ptr = _ptr;
+                int64_t *old_ptr = _ptr;
                 DumpInstr[*_ptr]();
                 _ptr = old_ptr;
             }
@@ -65,7 +65,7 @@ uint64_t Interpretator::Run(bool verbose) {
     }
 
     if (verbose) {
-        uint8_t *old_ptr = _ptr;
+        int64_t *old_ptr = _ptr;
         DumpInstr[*_ptr]();
         _ptr = old_ptr;
     }
@@ -82,7 +82,7 @@ void Interpretator::Dump() {
     }
 
     while (*_ptr != OPCODE_EXIT && *_ptr != OPCODE_NOP) {
-        uint8_t opcode = *_ptr;
+        int64_t opcode = *_ptr;
         if (opcode > OPCODE_EXIT || DumpInstr[opcode] == nullptr) {
             throw std::runtime_error("Invalid opcode: " + std::to_string(opcode));
         }
