+++ tests/test2.cpp
@@ -2,14 +2,14 @@
 
 int main() {
     Interpretator i;
-    i._bytecode = reinterpret_cast<uint8_t *>(calloc(10000, 1));
+    i._bytecode = reinterpret_cast<int64_t *>(calloc(10000, 1));
     i._bytecode_len = 100;
     i._ptr = i._bytecode;
     Frame f(3);
     i._frames.push(f);
-    uint8_t * ptr = i._bytecode;
-    // std::cout << "ptr: " << reinterpret_cast<uint64_t>(ptr) << std::endl;
-    // ptr_history.push_back(reinterpret_cast<uint64_t>(ptr));
+    int64_t * ptr = i._bytecode;
+    // std::cout << "ptr: " << reinterpret_cast<int64_t>(ptr) << std::endl;
+    // ptr_history.push_back(reinterpret_cast<int64_t>(ptr));
     i.create_lmov(1, 1);
     i.create_ladd(1, 1, 1);
     i.create_ladd(1, 1, 1);
